Rewrite for oneshot pub/sub
