+++ rumqttc/src/lib.rs
@@ -106,7 +106,7 @@ use std::time::Duration;
 mod client;
 mod eventloop;
 mod framed;
-mod new;
+mod oneshot;
 mod state;
 mod tls;
 
@@ -115,10 +115,10 @@ pub use client::{AsyncClient, Client, ClientError, Connection};
 pub use eventloop::{ConnectionError, Event, EventLoop};
 pub use mqttbytes::v4::*;
 pub use mqttbytes::*;
-pub use new::{ReqHandler, Publisher, Subscriber};
+pub use oneshot::{MqttHandler, Publisher, Subscriber};
 pub use state::{MqttState, StateError};
-pub use tokio_rustls::rustls::ClientConfig;
 pub use tls::Error;
+pub use tokio_rustls::rustls::ClientConfig;
 
 pub type Incoming = Packet;
 
@@ -439,7 +439,11 @@ impl MqttOptions {
     }
 
     /// Username and password
-    pub fn set_credentials<U: Into<String>, P: Into<String>>(&mut self, username: U, password: P) -> &mut Self {
+    pub fn set_credentials<U: Into<String>, P: Into<String>>(
+        &mut self,
+        username: U,
+        password: P,
+    ) -> &mut Self {
         self.credentials = Some((username.into(), password.into()));
         self
     }
@@ -673,7 +677,7 @@ impl std::convert::TryFrom<url::Url> for MqttOptions {
             inflight,
             last_will: None,
             conn_timeout,
-            manual_acks: false
+            manual_acks: false,
         })
     }
 }
